scraper/requirements.txt
# Wrestling Analytics Scraper Dependencies

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Database
supabase>=1.0.0
psycopg2-binary>=2.9.0

# API framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Data validation and serialization
pydantic>=2.5.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
hypothesis>=6.88.0

# Logging and monitoring
structlog>=23.2.0

# Utilities
python-dateutil>=2.8.0
tenacity>=8.2.0